requirements.txt
asgiref==3.11.1
cloudinary==1.34.0
dj-database-url==2.3.0
Django==6.0.2
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
razorpay==2.0.0
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.8.2
openpyxl==3.1.5


